Corrigindo caquinha que fiz na tela de perfil

+++ src/pages/Profile.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import { Button } from 'react-bootstrap';
 import { useHistory } from 'react-router-dom';
 
@@ -7,7 +7,14 @@ import Header from '../components/Header';
 
 export default function Profile() {
   const history = useHistory();
-  const { email } = JSON.parse(localStorage.getItem('user'));
+  const [userEmail, setEmail] = useState([]);
+
+  useEffect(() => {
+    if (localStorage.getItem('user')) {
+      const { email } = JSON.parse(localStorage.getItem('user'));
+      setEmail(email);
+    }
+  }, []);
 
   const clickButton = ({ target: { name } }) => {
     const domain = name === 'login' ? '' : name;
@@ -15,37 +22,39 @@ export default function Profile() {
     history.push(`/${domain}`);
   };
 
-  return (
-    <>
-      <Header />
-      <p data-testid="profile-email">
-        { email }
-      </p>
-      <Button
-        data-testid="profile-done-btn"
-        name="receitas-feitas"
-        onClick={ clickButton }
-        type="button"
-      >
-        Receitas Feitas
-      </Button>
-      <Button
-        data-testid="profile-favorite-btn"
-        type="button"
-        name="receitas-favoritas"
-        onClick={ clickButton }
-      >
-        Receitas Favoritas
-      </Button>
-      <Button
-        data-testid="profile-logout-btn"
-        type="button"
-        name="login"
-        onClick={ (e) => { clickButton(e); localStorage.clear(); } }
-      >
-        Sair
-      </Button>
-      <Footer />
-    </>
-  );
+  if (userEmail) {
+    return (
+      <>
+        <Header />
+        <p data-testid="profile-email">
+          { userEmail }
+        </p>
+        <Button
+          data-testid="profile-done-btn"
+          name="receitas-feitas"
+          onClick={ clickButton }
+          type="button"
+        >
+          Receitas Feitas
+        </Button>
+        <Button
+          data-testid="profile-favorite-btn"
+          type="button"
+          name="receitas-favoritas"
+          onClick={ clickButton }
+        >
+          Receitas Favoritas
+        </Button>
+        <Button
+          data-testid="profile-logout-btn"
+          type="button"
+          name="login"
+          onClick={ (e) => { clickButton(e); localStorage.clear(); } }
+        >
+          Sair
+        </Button>
+        <Footer />
+      </>
+    );
+  }
 }
